+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.1.0
+# 0.1.0-dev
 
 - **BREAKING** dart upgrade to 3.3.0